buyshop/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
cryptography==44.0.2
dj-database-url==2.3.0
Django==5.2
django-allauth==65.7.0
django-cors-headers==4.7.0
django-phonenumber-field==8.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-yasg==1.21.10
ecdsa==0.19.1
idna==3.10
inflection==0.5.1
msgpack==1.1.0
packaging==24.2
phonenumbers==9.0.2
pillow==11.1.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.0
python-jose==3.4.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rsa==4.9
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.0
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.1.1
urllib3==2.3.0
user-agents==2.2.0
